requirements.txt
Django==3.1.1
django_extensions==3.0.9
Pillow==7.2.0 # For imageField
django_guardian==2.3.0
django-debug-toolbar >= 3.1
django_silk==4.1.0
 django_treenode==0.15.0